Add ApiBaseController and optimize source code
