+++ SourceCode/13_Mar_2016/EBROnline.sln
@@ -9,8 +9,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EBROnline.Model", "EBROnlin
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EBROnline", "EBROnline\EBROnline.csproj", "{28F4E0A3-13CF-4690-A162-30B89231E23E}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EBROnlineAutomationTest", "EBROnlineAutomationTest\EBROnlineAutomationTest.csproj", "{4BFD400D-9E5A-42AA-9173-BCDC43EDD30A}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		CD_ROM|Any CPU = CD_ROM|Any CPU
@@ -50,16 +48,6 @@ Global
 		{28F4E0A3-13CF-4690-A162-30B89231E23E}.Release|Any CPU.Build.0 = Release|Any CPU
 		{28F4E0A3-13CF-4690-A162-30B89231E23E}.SingleImage|Any CPU.ActiveCfg = Release|Any CPU
 		{28F4E0A3-13CF-4690-A162-30B89231E23E}.SingleImage|Any CPU.Build.0 = Release|Any CPU
-		{4BFD400D-9E5A-42AA-9173-BCDC43EDD30A}.CD_ROM|Any CPU.ActiveCfg = Release|Any CPU
-		{4BFD400D-9E5A-42AA-9173-BCDC43EDD30A}.CD_ROM|Any CPU.Build.0 = Release|Any CPU
-		{4BFD400D-9E5A-42AA-9173-BCDC43EDD30A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{4BFD400D-9E5A-42AA-9173-BCDC43EDD30A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{4BFD400D-9E5A-42AA-9173-BCDC43EDD30A}.DVD-5|Any CPU.ActiveCfg = Debug|Any CPU
-		{4BFD400D-9E5A-42AA-9173-BCDC43EDD30A}.DVD-5|Any CPU.Build.0 = Debug|Any CPU
-		{4BFD400D-9E5A-42AA-9173-BCDC43EDD30A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{4BFD400D-9E5A-42AA-9173-BCDC43EDD30A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{4BFD400D-9E5A-42AA-9173-BCDC43EDD30A}.SingleImage|Any CPU.ActiveCfg = Release|Any CPU
-		{4BFD400D-9E5A-42AA-9173-BCDC43EDD30A}.SingleImage|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
